requirements.txt
PyQt5>=5.15
pycryptodome>=3.19
